requirements.txt
async-lru==1.0.2
pyaes==1.6.1
Pyrogram==1.2.9
PySocks==1.7.1
python-dotenv==0.19.0
schedule==1.1.0
